#51 Refactor dto and enums
