+++ package.json
@@ -76,7 +76,8 @@
       "^.+\\.(t|j)s$": "ts-jest"
     },
     "collectCoverageFrom": [
-      "**/*.(t|j)s"
+      "**/*.(t|j)s",
+      "!migrations/**"
     ],
     "coverageDirectory": "../coverage",
     "testEnvironment": "node"
