+++ src/customer/customer.controller.ts
@@ -3,8 +3,8 @@ import { Get, Post, Put, Body } from '@nestjs/common'
 import { Query, Param, ParseIntPipe } from '@nestjs/common'
 import { NotFoundException } from '@nestjs/common'
 
-import { PaginationDto } from '../dto/pagination.dto'
-import { PaginatedResultDto } from '../dto/paginated-result.dto'
+import { PaginationDto } from '../common/dto/pagination.dto'
+import { PaginatedResultDto } from '../common/dto/paginated-result.dto'
 import { CreateCustomerDto } from './dto/create-customer.dto'
 import { UpdateCustomerDto } from './dto/update-customer.dto'
 import { CustomerService } from './customer.service'
