+++ src/order/dto/create-order.dto.ts
@@ -1,5 +1,5 @@
 import { IsNotEmpty, IsNumber, IsString } from 'class-validator'
-import { OrderStatus } from '../enums/order-status.enum'
+import { OrderStatus } from '../../common/enums/order-status.enum'
 
 export class CreateOrderDto {
   @IsNumber()
